+++ NEWS.md
@@ -2,8 +2,10 @@
 
 * A new argument `regex` has been added to the `docs()`-method. It can be used to limit the analysis to documents matched by the regular expression.
 * A bug has been removed from the `cooccurrences()`-method that resulted in an overestimation of the cooccurrence of topics when topics had been renumbered.
-
-
+* A new field `topics` has been introduced in the `Topicanalysis` class. A topics-document-matrix can #
+be stored here. As retrieving the topics-document-matrix can be slow, subsequent operations on this matrix are faster now.
+* The `cooccurrencs()`-method for the `TopicModel` class now has an argument `topic_matrix`, so that a topic-document-matrix prepared outside the method can be re-used. May be used for speeding up operations.
+* Added new argument `verbose` to `Topicanalysis$cooccurrences()`-method, which is passed into S4 method `cooccurrences()`.
 
 ## v0.1.14
   * The preliminary Rmarkdown template to use 'gensim' for topic modelling has been turned into a (somewhat) consolidated Rmarkdown template.